fix: model annotations as nested objects
Updates all Content types to represent annotations as nested objects
instead of directly on the content itself. The existing ctors have
been moved to deprecated shim ctors for backwards-compat.

This brings the implementation in line with the content types as
modeled by the specification. An integration test has been added
to ensure this doesn't drift in the future.

https://github.com/modelcontextprotocol/modelcontextprotocol/blob/c87a0da6d8c2436d56a6398023c80b0562224454/schema/2025-03-26/schema.json#L1970-L1973