Support server to client notifications from the stateless transport
The MCP spec allows stateless servers to send notifications to the
client during a request. The response needs to be upgraded to SSE
and the notifications are send in a stream until the final result
is sent.

This commit adds a `sendNotification` method to the transport
context allowing each transport implementation to implement it
or not. In this commit, HttpServletStatelessServerTransport
implements the method and when the caller first sends a
notification, the response is changed to `TEXT_EVENT_STREAM`
and events are then streamed until the final result.

This change will allow future features such as logging, list
changes, etc. should we ever decide to support sessions in
some manner. Even if we don't support sessions, sending progress
notifications is a useful feature by itself.